models/python-server/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
transformers==4.27.4
mediapipe==0.10.0
tensorflow==2.12.0
python-dotenv==1.0.0